Refactoring, added custom exceptions, error handling, changed jobs selection logic when filtering by skills, refactor get_job_by_id logic

+++ server/services/user.py
@@ -1,11 +1,11 @@
 import bcrypt
-from fastapi import Depends, HTTPException, status
+from fastapi import Depends
 from typing import Annotated
 
-from models import User
 from schemas import UserAuthSchema, UserSchema
 from repositories import UserRepositoryDependency
 from utils.access_token import create_access_token, verify_access_token
+from exception import UserNotFoundError, AccessDeniedError, UnauthorizedError
 
 class UserService:
     def __init__(self, user_repository: UserRepositoryDependency): 
@@ -23,21 +23,20 @@ class UserService:
             "email" : new_user.email
         }
 
-        access_token = create_access_token(token_data)
-        return access_token
+        return create_access_token(token_data)
     
     def login_user(self, user_auth: UserAuthSchema) -> str:
         user = self.repository.get_user_by_email(user_auth.email)
         
         if user is None:
-            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid email or password")
+            raise UnauthorizedError(message="Invalid email or password!")
 
         password_check = bcrypt.checkpw(
             user_auth.password.encode("utf-8"), user.password.encode("utf-8")
         )
 
         if not password_check:
-            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Invalid email or password")
+            raise UnauthorizedError(message="Invalid email or password!")
         
         token_data = {
             "id" : user.id,
@@ -49,21 +48,27 @@ class UserService:
 
     def update_token (self, token: str | None) -> str:
         decoded = verify_access_token(token)
-        new_access_token = create_access_token(decoded)
-        return new_access_token
+        new_token = {
+            "id" : decoded["id"],
+            "email" : decoded["email"]
+        }
+         
+        return create_access_token(new_token)
             
     def get_user(self, id: int) -> UserSchema:
         user = self.repository.get_user_by_id(id)        
         if user is None:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="This user was not found")
-        
+            raise UserNotFoundError(id)
         return UserSchema.model_validate(user)
     
-    def delete_user(self, id: int) -> str:
+    def delete_user(self, id: int, token: str) -> str:
+        decoded = verify_access_token(token)
+        if decoded["id"] != id:
+            raise AccessDeniedError()
+    
         user = self.repository.get_user_by_id(id)
-
         if user is None:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="This user was not found")
+            raise UserNotFoundError(id)
         
         self.repository.delete_user(user)
         return "User was deleted"
